+++ libs/common/src/key-management/crypto/abstractions/encrypt.service.ts
@@ -1,9 +1,9 @@
-import { Decryptable } from "../../../platform/interfaces/decryptable.interface";
-import { Encrypted } from "../../../platform/interfaces/encrypted";
-import { InitializerMetadata } from "../../../platform/interfaces/initializer-metadata.interface";
-import { EncArrayBuffer } from "../../../platform/models/domain/enc-array-buffer";
-import { EncString } from "../../../platform/models/domain/enc-string";
-import { SymmetricCryptoKey } from "../../../platform/models/domain/symmetric-crypto-key";
+import { Decryptable } from "@bitwarden/common/platform/interfaces/decryptable.interface";
+import { Encrypted } from "@bitwarden/common/platform/interfaces/encrypted";
+import { InitializerMetadata } from "@bitwarden/common/platform/interfaces/initializer-metadata.interface";
+import { EncArrayBuffer } from "@bitwarden/common/platform/models/domain/enc-array-buffer";
+import { EncString } from "@bitwarden/common/platform/models/domain/enc-string";
+import { SymmetricCryptoKey } from "@bitwarden/common/platform/models/domain/symmetric-crypto-key";
 
 export abstract class EncryptService {
   abstract encrypt(plainValue: string | Uint8Array, key: SymmetricCryptoKey): Promise<EncString>;
